% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable.R
\name{handsontableOutput}
\alias{handsontableOutput}
\title{Create a Handsontable output element}
\usage{
handsontableOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{The output identifier for the widget}

\item{width}{The width of the widget}

\item{height}{The height of the widget}
}
\value{
A Shiny output element
}
\description{
Create a Handsontable output element
}
