% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsontable.R
\name{renderHandsontable}
\alias{renderHandsontable}
\title{Render a Handsontable widget}
\usage{
renderHandsontable(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a Handsontable widget}

\item{env}{The environment in which to evaluate expr}

\item{quoted}{Is expr a quoted expression?}
}
\value{
A Shiny render function
}
\description{
Render a Handsontable widget
}
